count frequency in an array
given an int array, return the count of occurances of the most frequent number in the array.

The solution is not the most compact, nor is it the optimal if speed is considered. It is a practice of maps.

TreeMap was used, but it could be any map, since the specifics of TreeMap could not be used here regretfully.